Bug-fix. На данным момент решение набираее 1774 балла. Виной всему - ложные срабатывания. Постараюсь от них избавиться

+++ Hola/source/Hola/Code/Analyze/CodeAnalyzer.cs
@@ -24,7 +24,7 @@ namespace Hola.Code.Analyze
             Code = code;
             Language = language;
         }
-        public virtual double Compare(CodeAnalyzer code)
+        public virtual decimal Compare(CodeAnalyzer code)
         {
             if (code.Code == Code) return 1;
             else return 0;
